requirements.txt
Django
djangorestframework
docker
docker-compose
drf-yasg
environs
psycopg2
